test_client/requirements.txt
Appium-Python-Client==0.47
behave==1.2.6
selenium==3.141.0
